refactor(service): separate functionalities in CalculatorService into distinct objects for better OOP structure
