+++ src/main/java/calculator/exception/InvalidInputException.java
@@ -1,0 +1,7 @@
+package calculator.exception;
+
+public class InvalidInputException extends IllegalArgumentException{
+    public InvalidInputException(String message) {
+        super(message);
+    }
+}
