+++ src/main/java/calculator/exception/NegativeNumberException.java
@@ -1,0 +1,7 @@
+package calculator.exception;
+
+public class NegativeNumberException extends IllegalArgumentException{
+    public NegativeNumberException(String message) {
+        super(message);
+    }
+}
