+++ src/main/java/calculator/model/Parser.java
@@ -1,0 +1,4 @@
+package calculator.model;
+
+public class Parser {
+}
