+++ src/main/java/calculator/model/SumCalculator.java
@@ -1,0 +1,4 @@
+package calculator.model;
+
+public class SumCalculator {
+}
