+++ src/main/java/calculator/validator/InputValidator.java
@@ -1,0 +1,10 @@
+package calculator.validator;
+
+public class InputValidator {
+
+    public int inputValidate(String input) {
+        if (input == null || input.isEmpty()) {
+            return 0;
+        }
+    }
+}
